# Deduplicated dequantization code