@ggerganov 2023-05-14
I don't see performance degradation on RTX 4080